Rename file to `generic_properties_treeview.py`
