+++ examples/python/gui_demo/components/generic_properties_treeview.py
